Retarget perception.memory imports after memory2 was renamed to memory.
